// src/app/core/models/persona.ts
export interface Persona{
  idPersona: number;
  identificacion: number;
  nombre: string;
  usuarioLogin: string;
  telefono: number;
  email: string;
  idTipoCategoriaPersona: number;
  idTipoRegimen: number;
  idTipoArea: number;
  idTipoCargo: number;
  idTipoEstadoPersona: number;
  idTipoRol: number;
  idUsuario: number;
}